[REF] point_of_sale: reorganize assets
This commits simplifies the asset management in pos. It also opens
the possibility to reduce the size of assets loaded in POS, see
the contents of `point_of_sale.pos_assets_backend`. Looking at that
bundle, it's now obvious that we can cherry pick the list of assets
we need from `web.assets_common` and `web.assets_backend`.

Note that manual picking the `web` assets is not done in this commit.
