+++ addons/point_of_sale/static/src/backend/debug_manager.js
