+++ addons/point_of_sale/static/src/backend/tours/point_of_sale.js
