+++ addons/point_of_sale/static/src/backend/web_overrides/pos_config_form.js
