+++ addons/point_of_sale/static/src/entry/chrome_adapter.js
