+++ addons/point_of_sale/static/src/entry/main.js
@@ -2,7 +2,7 @@
 
 import { startWebClient } from "@web/start";
 
-import { ChromeAdapter } from "@point_of_sale/js/chrome_adapter";
+import { ChromeAdapter } from "@point_of_sale/entry/chrome_adapter";
 import Registries from "point_of_sale.Registries";
 import { registry } from "@web/core/registry";
