+++ addons/point_of_sale/views/pos_assets_common.xml
@@ -4,7 +4,6 @@
 <template id="point_of_sale.assets_common" name="POS Assets Common">
     <t t-call-assets="point_of_sale.pos_assets_backend_style" t-js="false"/>
     <t t-call-assets="point_of_sale.assets" t-js="false" />
-    <t t-call-assets="web.assets_common" t-css="false" />
     <t t-call-assets="point_of_sale.pos_assets_backend" t-css="false" />
     <t t-call-assets="point_of_sale.assets" t-css="false" />
 </template>
