+++ addons/point_of_sale/views/pos_assets_qunit.xml
@@ -2,7 +2,7 @@
 <odoo>
 
 <template id="point_of_sale.qunit_suite_assets">
-    <t t-call-assets="point_of_sale.pos_assets_backend" t-js="false"/>
+    <t t-call-assets="web.assets_backend" t-js="false" />
     <t t-call-assets="point_of_sale.tests_assets" t-js="false"/>
     <t t-call-assets="point_of_sale.tests_assets" t-css="false"/>
     <style>
